shuffle some files around to make structure a bit neater
