+++ openapi-codegen/core/src/main/scala/sttp/tapir/codegen/dedup/PackageReuseContext.scala
@@ -1,6 +1,8 @@
-package sttp.tapir.codegen
+package sttp.tapir.codegen.dedup
 
 import sttp.tapir.codegen.openapi.models.OpenapiModels.OpenapiDocument
+import sttp.tapir.codegen.dedup.PackageReuseContext
+import sttp.tapir.codegen.security.SecurityWrapperDefn
 
 /** Describes a dependency package whose models may be reused via type aliases. */
 case class PackageReuseContext(
@@ -46,4 +48,47 @@ object PackageReuseContext {
        |val $name = ${ctx.dependencyModelPath}.$name""".stripMargin
 
   def isReusedSchema(name: String, ctx: PackageReuseContext): Boolean = ctx.reusedSchemas.contains(name)
+}
+
+
+object GenerationMeta {
+  val default: GenerationMeta = GenerationMeta(Seq("TapirGeneratedEndpointsSchemas"), false, false, Nil, Set.empty, Nil, 0, Set.empty, Set.empty)
+}
+case class GenerationMeta(
+                           schemaFiles: Seq[String],
+                           hasValidators: Boolean,
+                           schemasContainAny: Boolean,
+                           explicitNonObjTypes: Seq[String],
+                           security: Set[SecurityWrapperDefn],
+                           extensions: Seq[(String, String, String)],
+                           schemaObjectCount: Int,
+                           allTransitiveJsonParamRefs: Set[String],
+                           jsonParamRefs: Set[String],
+                         ) {
+  def partition(securityWrappers: Set[SecurityWrapperDefn]): (Set[SecurityWrapperDefn], Set[SecurityWrapperDefn]) = {
+    val changed = scala.collection.mutable.Set.empty[SecurityWrapperDefn]
+    val m = scala.collection.mutable.Set.empty[SecurityWrapperDefn]
+    val ct = scala.collection.mutable.Set.empty[String]
+    securityWrappers.foreach(w =>
+      if (!security.contains(w)) {
+        changed += w
+        w.schemas.map(_.typeName).foreach(t => ct += t)
+      } else m += w
+    )
+    def checkChanged: Unit = {
+      var changedCount = 0
+      val mSnapshot = m.toSet
+      mSnapshot.foreach(w =>
+        if (w.schemas.map(_.typeName).exists(ct.contains)) {
+          changedCount += 1
+          w.schemas.map(_.typeName).foreach(t => ct += t)
+          changed += w
+          m.remove(w)
+        }
+      )
+      if (changedCount != 0) checkChanged
+    }
+    checkChanged
+    m.toSet -> changed.toSet
+  }
 }
