+++ openapi-codegen/core/src/main/scala/sttp/tapir/codegen/json/JsonSerdeGenerator.scala
@@ -23,7 +23,7 @@ import sttp.tapir.codegen.openapi.models.OpenapiSchemaType.{
   OpenapiSchemaStringType,
   OpenapiSchemaUUID
 }
-import sttp.tapir.codegen.PackageReuseContext
+import sttp.tapir.codegen.dedup.PackageReuseContext
 import sttp.tapir.codegen.util.NameHelpers.{addName, indent, uncapitalise}
 
 import scala.annotation.tailrec
