+++ openapi-codegen/core/src/test/scala/sttp/tapir/codegen/PackageReuseSpec.scala
@@ -2,17 +2,9 @@ package sttp.tapir.codegen
 
 import org.scalatest.flatspec.AnyFlatSpec
 import org.scalatest.matchers.should.Matchers
+import sttp.tapir.codegen.dedup.{GenerationMeta, PackageReuseContext, SchemaComparer}
 import sttp.tapir.codegen.openapi.models.OpenapiSchemaType
-import sttp.tapir.codegen.openapi.models.OpenapiSchemaType.{
-  NumericRestrictions,
-  OpenapiSchemaConstantString,
-  OpenapiSchemaEnum,
-  OpenapiSchemaField,
-  OpenapiSchemaInt,
-  OpenapiSchemaObject,
-  OpenapiSchemaRef,
-  OpenapiSchemaString
-}
+import sttp.tapir.codegen.openapi.models.OpenapiSchemaType.{NumericRestrictions, OpenapiSchemaConstantString, OpenapiSchemaEnum, OpenapiSchemaField, OpenapiSchemaInt, OpenapiSchemaObject, OpenapiSchemaRef, OpenapiSchemaString}
 import sttp.tapir.codegen.openapi.models.OpenapiModels.OpenapiDocument
 import sttp.tapir.codegen.openapi.models.OpenapiComponent
 import sttp.tapir.codegen.openapi.models.OpenapiModels.OpenapiInfo
@@ -79,7 +71,7 @@ class PackageReuseContextSpec extends AnyFlatSpec with Matchers {
       petDoc(),
       "sttp.tapir.generated.core",
       "TapirGeneratedEndpoints",
-      GenerationMeta.default,
+      GenerationMeta.default
     )
     ctx.reusedSchemas shouldBe Set("Pet")
     ctx.dependencyModelPath shouldBe "sttp.tapir.generated.core.TapirGeneratedEndpoints"
